+++ src/app/contact/page.tsx
@@ -1,6 +1,6 @@
 'use client'; // This marks the entire file as a Client Component
-import CloudBackground from '@/common/backgrounds/cloud/CloudBackground';
-import { MenuModel } from '@/common/Menu/MenuModel';
+import CloudBackground from '@/components/backgrounds/cloud/CloudBackground';
+import { MenuModel } from '@/components/Menu/MenuModel';
 
 export default function Contact() {
     return (
