+++ src/app/login/page.tsx
@@ -1,7 +1,7 @@
 'use client'; // This marks the entire file as a Client Component
 import { XYLabelModel } from '@/types';
-import XYButton from '@/common/Inputs/XYButton';
-import { MenuModel } from '@/common/Menu/MenuModel';
+import XYButton from '@/components/Inputs/XYButton';
+import { MenuModel } from '@/components/Menu/MenuModel';
 import { PrimeReactProvider } from 'primereact/api';
 import { confirmDialog } from 'primereact/confirmdialog';
 import { Password } from 'primereact/password';
