+++ src/app/new/page.tsx
@@ -1,0 +1,16 @@
+'use client';
+import React from 'react';
+
+export const New = () => {
+    return (
+        <div className='h-full w-full'>
+            <h1>Welcome to XYS</h1>
+            <p>
+                Please complete the short form below and a member of our
+                onboarding team will reach out to you within 1-4 buisness days
+                with everything you will need to know and more.
+            </p>
+            {/* form goes here  */}
+        </div>
+    );
+};
