+++ src/app/team/page.tsx
@@ -1,8 +1,8 @@
 'use client'; // This marks the entire file as a Client Component
-import { MenuModel } from '@/common/Menu/MenuModel';
+import { MenuModel } from '@/components/Menu/MenuModel';
 import { NextPage } from 'next';
 import OurTeam from './components/OurTeam';
-import { XYSToTop } from '@/common';
+import { XYSToTop } from '@/components';
 
 interface Props {}
