+++ src/components/backgrounds/cloud/CloudBackground.tsx
