+++ src/components/backgrounds/index.ts
