+++ src/components/backgrounds/svg/SVGBackground.tsx
