+++ src/components/backgrounds/svg/bck.scss
