+++ src/types/teamTypes.ts
@@ -11,5 +11,5 @@ export interface AboutUsModel {
     id: number;
     text: string;
     type: XYType;
-    imageSrc : string;
+    imageSrc: string;
 }
